Graph tooltip, display and interval enhancements
At a high level, changes as follows:
1. Change tooltip to show give context related info, e.g. for a daily stat only show YMD
2. Add support for touch devices within flot
3. Change the Interval to display and accept human readable figures, e.g. 1d, 2w, 4h
4. Set the minTickSize to align with the YMD intervals
5. bar position is placed to the right for 'delta' to help show what time duration this covers
6. Store the requesttype, e.g. fixed, daily, weekly ...
7. Align the start/ end dates to show the entire bar

embed.php:
Added some padding to the navigation bar to prevent the border appearing outside of the specified window

view.php:
Bring the datepicker inputs into one block
Assign values for the request-type field to enable setting of this field when graph config is reloaded

vis.helper.js:
Added 2 vars graphDisplayOffsetStart and graphDisplayOffsetEnd. This was triggered because for a graph with 'weekly' deltas the display considers the start of week to be a Thursday (Jan 1st 1970 was a Thursday), but the get_data_DMY PHPFIna function aligns weeks to a Monday, meaning the display would be out of alignment with the data. These offsets compensate for this.
This could be further expanded to support the user choosing which day is the start of the week or otherwise for monthly, which should be the 'start' of a month - maybe to align with when you statement is produced!?
panright() and panleft() have been joined into a pan() function, with a parameter of direction and the request-type.
1. If request-type is monthly, then change start and end dates to align with 1st month multiples
2. Changed the shiftsize to be the number of intervals and then apply a 20% delta rounded to the nearest interval size to keep start and end dates equivalent
calc_interval() change to:
1. Reorder to change interval if not fixed first
2. Then align start and end dates to the specified/ calculated interval
3. Where 'monthly' align start/ end dates to 1st of the month
4. setting up of graphDisplayOffsetStart and graphDisplayOffsetEnd to display weekly aligned to a monday
tooltip() added parameter for parent object to enable aligning of the tooltip within that object scope

graph.js:
Replace panleft/ panright with pan() and the request-type
Change tooltip display logic to change whether the data or series indexes change, otherwise stacked bars won't display other items when stacked.
Changed tooltip content to display times and dates as follows:
1. If a daily interval just the YMD
2. If interval >= daily show start to end for interval as start date to end date
3. Otherwise show start to end of interval as start date and time to end date and time
NOTE: Logic forces this to Locale en-GB, I need to work on making this pick up the browser's settings instead.
Added code to support touch device interaction with the flot graph
Added convertReadableToSecs(readable) and convertSecsToReadable(secs) to convert the interval from seconds into a human readable form, i.e. seconds are displayed as a number followed by a letter, which break the interval down into Year, Month, Week, Day, Hour, Second.
NOTE: I've left Minute out as M is already used for Month, but am happy to rename if appropriate.
NOTE: It would be nice to make this a little more international, hence if we can have an equivalent set of characters, then we can potentially support difference locales.
To reduce changes to the stored data and processes, view.interval is still used by all other aspects of the code, with the human readable stored as displayInterval and this variable is shown within the setup -> graph page.
If yaxis-min and yaxis-max are left blank replace changed blank text to 'auto'
graph_reload() changed the ajax call to extend the data collection range to align with the weekly Monday vs Thursday offset. Also offloaded the start/ end times to view.calc_interval() to provide a single source of date change and management for monthly etc
graph_draw() changes:
1. set the minTickSize to align with the interval size, works well for monthly and daily, but weekly alternates between 7 and 8 day intervals on the x-axis, meaning the ticks don't align very well.
2. Fixed the stacked line graph fill as I'd mistakenly place fill twice within the definition with my previous change
3. Where bars are used WITHOUT deltas the bar is centred to the data point, hence the start of the graph starts 1/2 and interval earlier to show the entire bar width
4. When displaying the feed list add a blank before the "move down" arrow to align all down arrows on the right, hopefully make it easier to follow for a new user.
On graph-select change:
1. set axis min/ max to auto if blank
2. set access and contents from the stored fixinterval and requesttype
3. load requesttype (to remember the annual/ monthly/ weekly/ daily settings for a save graph
4. forced interval to a Number to support use of === elsewhere in the code, originally the number was stored in the DB as a string